requirements.txt
beautifulsoup4==4.10.0
numpy==1.21.1
pandas==1.4.0
python-dotenv==0.19.2
PyYAML==6.0
argparse==1.4.0
requests==2.27.1
vk-requests==1.2.1
